move grpc definition into dock package from the grpc package
