+++ pkg/controller/policy/executor/executorintervalsnapshot.go
@@ -28,8 +28,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/opensds/opensds/pkg/grpc/dock/client"
-	pb "github.com/opensds/opensds/pkg/grpc/opensds"
+	"github.com/opensds/opensds/pkg/dock/client"
+	pb "github.com/opensds/opensds/pkg/dock/proto"
 	api "github.com/opensds/opensds/pkg/model"
 	"golang.org/x/net/context"
 )
