+++ pkg/controller/volume/volumecontroller.go
@@ -24,8 +24,8 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/opensds/opensds/pkg/grpc/dock/client"
-	pb "github.com/opensds/opensds/pkg/grpc/opensds"
+	"github.com/opensds/opensds/pkg/dock/client"
+	pb "github.com/opensds/opensds/pkg/dock/proto"
 	api "github.com/opensds/opensds/pkg/model"
 	"golang.org/x/net/context"
 )
