+++ pkg/controller/volume/volumecontroller_test.go
@@ -19,8 +19,8 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/opensds/opensds/pkg/grpc/dock/client"
-	pb "github.com/opensds/opensds/pkg/grpc/opensds"
+	"github.com/opensds/opensds/pkg/dock/client"
+	pb "github.com/opensds/opensds/pkg/dock/proto"
 	api "github.com/opensds/opensds/pkg/model"
 	"golang.org/x/net/context"
 	"google.golang.org/grpc"
