+++ pkg/dock/proto/dock.proto
@@ -29,7 +29,7 @@
 
 syntax = "proto3";
 
-package opensds;
+package proto;
 
 // The dock service definition.
 service Dock {
